# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: HASN Digital Services Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register with phone and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with phone and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized

  # Users (Admin)
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders (mine or all for admin)
      parameters:
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: status
          schema:
            type: string
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderList"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # Banners
...
  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get site settings
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteSettings"
    patch:
      operationId: updateSettings
      tags: [settings]
      summary: Update site settings (super admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SiteSettingsUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteSettings"

  # Admin stats
...
  /admin/audit-logs:
    get:
      operationId: listAuditLogs
      tags: [admin]
      summary: List audit log entries
      parameters:
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: Audit logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditLogList"

  # Payment methods
...
components:
  schemas:
...
    RegisterInput:
      type: object
      required: [phone, password, name]
      properties:
        phone:
          type: string
        password:
          type: string
        name:
          type: string
    LoginInput:
      type: object
      required: [phone, password]
      properties:
        phone:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/User"

    # User
    User:
      type: object
      required: [id, phone, name, role, isBlocked, walletBalance, createdAt]
      properties:
        id:
          type: integer
        phone:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [user, admin, super_admin]
        isBlocked:
          type: boolean
        walletBalance:
          type: number
        createdAt:
          type: string
...
    Order:
      type: object
      required: [id, userId, serviceId, serviceName, amount, status, quantity, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        serviceId:
          type: integer
        serviceName:
          type: string
        serviceImage:
          type: ["string", "null"]
        amount:
          type: number
        quantity:
          type: integer
        status:
          type: string
          enum: [pending, processing, completed, failed, cancelled]
        targetId:
          type: ["string", "null"]
        providerOrderId:
          type: ["string", "null"]
        createdAt:
          type: string
    OrderList:
      type: object
      required: [orders, total, page, limit]
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    OrderInput:
      type: object
      required: [serviceId, quantity, targetId]
      properties:
        serviceId:
          type: integer
        quantity:
          type: integer
        targetId:
          type: string

    # Upload
...
    SiteSettings:
      type: object
      required: [siteName]
      properties:
        siteName:
          type: string
        logoUrl:
          type: ["string", "null"]
        faviconUrl:
          type: ["string", "null"]
        instagramUrl:
          type: ["string", "null"]
        whatsappUrl:
          type: ["string", "null"]
        facebookUrl:
          type: ["string", "null"]
        telegramUrl:
          type: ["string", "null"]
        globalProfitMargin:
          type: ["number", "null"]
        maintenanceMode:
          type: boolean
        currency:
          type: string
    SiteSettingsUpdate:
      type: object
      properties:
        siteName:
          type: string
        logoUrl:
          type: string
        faviconUrl:
          type: string
        instagramUrl:
          type: string
        whatsappUrl:
          type: string
        facebookUrl:
          type: string
        telegramUrl:
          type: string
        globalProfitMargin:
          type: number
        maintenanceMode:
          type: boolean
        currency:
          type: string
...
    AuditLog:
      type: object
      required: [id, action, userId, createdAt]
      properties:
        id:
          type: integer
        action:
          type: string
        userId:
          type: ["integer", "null"]
        userName:
          type: ["string", "null"]
        details:
          type: ["string", "null"]
        createdAt:
          type: string
    AuditLogList:
      type: object
      required: [logs, total, page, limit]
      properties:
        logs:
          type: array
          items:
            $ref: "#/components/schemas/AuditLog"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer